This step is creating the prefabs and adding them to the other prefabs.
Three different explosion prefabs, differ only in size (sharing one Animator and Animations).

Created a couple of simple scripts:

SimpleSelfDestruction just calls a death timer with a delay immediately on Start. It is used to clean up animated effects when they are done playing.

SimpleCollider2D (added a few commits ago) is used to spawn an explosion effect when something collides with it. For now this is being used for all bullets and NPCs.

This is a good time to start iterating and making a few variations to test before you get too far ahead coding.

I also removed the NPCs from the scene and replaced them with different spawners for different enemies, so each enemy will get spawned on its own timer, with the simplest NPC most frequent and the biggest the least frequent.