Import assets for portraits. Using a frame and portrait set I found on Itch.
Lots of stuff to talk about in the frames regarding layout of sprites, padding and offset, 8 sided sprites and more.
I am also doing a scaling thing on the Image that brings the size of the corners and edge up, however because there is such a large gutter around the frames when I imported it, I had to add a negative offset to the frame (even though they are scaled to 100% of the parent!) because I was too lazy to properly reimport them (since I would have to manually set up all the 8 sprites again). Good enough for now, lots to discuss about UI layout and sprites, and something to fix later.